requirements.txt
Django==5.1
djangorestframework==3.15.0
django-cors-headers==4.3.1
channels==4.0.0
Pillow==11.2.1
django-extensions==3.2.3
django-filter==24.1
djangorestframework-simplejwt==5.3.1
daphne==4.0.0
scikit-learn==1.7.0
pandas==2.3.0
numpy==2.3.0
joblib==1.5.1
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==1.2.0
# psycopg2-binary==2.9.9  # غير متوافق مع Python 3.13
psycopg==3.1.18
nltk==3.8.1
# Alternative to implicit for collaborative filtering
scipy==1.14.1
# implicit==0.4.8  # Disabled due to Python 3.13 compatibility issues

# ./render_start.sh
# ./render_build.sh